feat: Adjusting in end points of projetos (#16)

+++ api/endpoints/projeto.py
@@ -1,11 +1,7 @@
 from typing import List
-
 from fastapi import APIRouter, status, Depends, HTTPException, Response
-
 from sqlalchemy.future import  select
-
 from sqlalchemy.orm import Session
-
 from models.projeto import Projeto
 from models.usuario import Usuario
 from schemas.projeto_schema import ProjetoSchema,ProjetoSchemaCreate,ProjetoSchemaUp, ProjetoSchemaAddMember
@@ -65,28 +61,29 @@ async def get_projeto(projeto_id: str, db: Session = Depends(get_session)):
 #PUT Projeto
 @router.put('/{projeto_id}', response_model=ProjetoSchema, status_code=status.HTTP_202_ACCEPTED)
 async def put_projeto(projeto_id: str, projeto: ProjetoSchemaUp, db: Session = Depends(get_session), usuario_logado: Usuario = Depends(get_current_user)):
-    query = select(Projeto).filter(Projeto.id == projeto_id)
-    result = db.execute(query)
-    projeto_up: Projeto = result.scalars().unique().one_or_none()
+    if usuario_logado:
+        query = select(Projeto).filter(Projeto.id == projeto_id)
+        result = db.execute(query)
+        projeto_up: Projeto = result.scalars().unique().one_or_none()
 
-    if projeto_up:
-        if projeto.titulo:
-            projeto_up.titulo = projeto.titulo
-        if projeto.descricao:
-            projeto_up.descricao = projeto.descricao
-            
-        projeto_up.data_up = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+        if projeto_up:
+            if projeto.titulo:
+                projeto_up.titulo = projeto.titulo
+            if projeto.descricao:
+                projeto_up.descricao = projeto.descricao
+                
+            projeto_up.data_up = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
 
-        db.commit()
-        return projeto_up
-    
-    else:
-        raise HTTPException(detail="Projeto não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
+            db.commit()
+            return projeto_up
+        
+        else:
+            raise HTTPException(detail="Projeto não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
         
 #DELETE Projeto
 @router.delete('/{projeto_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_projeto(projeto_id: str, db: Session = Depends(get_session), usuario_logado: Usuario = Depends(get_current_user)):
-    query = select(Projeto).filter(Projeto.id == projeto_id)
+    query = select(Projeto).filter(Projeto.id == projeto_id).filter(Projeto.autor_id == usuario_logado.id)
     result = db.execute(query)
     projeto_del: Projeto = result.scalars().unique().one_or_none()
 
@@ -104,47 +101,47 @@ async def delete_projeto(projeto_id: str, db: Session = Depends(get_session), us
 ##POST member in laboratory
 @router.post('/addMember/{projeto_id}/{usuario_id}', status_code=status.HTTP_201_CREATED)
 async def post_member(projeto_id: str , usuario_id: str , db: Session = Depends(get_session), usuario_logado: Usuario = Depends(get_current_user)): 
-   
-    query = select(Projeto).filter(Projeto.id == projeto_id).filter(Projeto.coordenador_id == usuario_logado.id)
-    result = db.execute(query)
-    projeto:Projeto = result.scalars().unique().one_or_none()
+    if usuario_logado:
+        query = select(Projeto).filter(Projeto.id == projeto_id)
+        result = db.execute(query)
+        projeto:Projeto = result.scalars().unique().one_or_none()
 
-    if projeto is None:
-        raise HTTPException(detail="Projeto não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
+        if projeto is None:
+            raise HTTPException(detail="Projeto não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
 
-    query = select(Usuario).filter(Usuario.id == usuario_id)
-    result = db.execute(query)
-    usuario: Usuario = result.scalars().unique().one_or_none()
+        query = select(Usuario).filter(Usuario.id == usuario_id)
+        result = db.execute(query)
+        usuario: Usuario = result.scalars().unique().one_or_none()
 
-    if usuario is None:
-        raise HTTPException(detail="Usuario não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
-    else:
-        if usuario in projeto.membros:
-            raise HTTPException(detail="Este usuário já é membro desse Projeto!", status_code=status.HTTP_404_NOT_FOUND)
+        if usuario is None:
+            raise HTTPException(detail="Usuario não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
         else:
-            projeto.membros.append(usuario)
+            if usuario in projeto.membros:
+                raise HTTPException(detail="Este usuário já é membro desse Projeto!", status_code=status.HTTP_404_NOT_FOUND)
+            else:
+                projeto.membros.append(usuario)
 
-            db.add(projeto)
-            db.commit()
-            return HTTPException(detail="Membro adicionado com sucesso com sucesso!", status_code=status.HTTP_201_CREATED)
+                db.add(projeto)
+                db.commit()
+                return HTTPException(detail="Membro adicionado com sucesso com sucesso!", status_code=status.HTTP_201_CREATED)
 
     
 #DELETE laboratorio
 @router.delete('/removeMember/{projeto_id}/{member_id}', status_code=status.HTTP_204_NO_CONTENT)
 async def delete_member_project(projeto_id: str, member_id: str, db: Session = Depends(get_session), usuario_logado: Usuario = Depends(get_current_user)):
-   
-    query = select(Projeto).filter(Projeto.id == projeto_id).filter(projeto.autor_id == usuario_logado.id)
-    result = db.execute(query)
-    projeto: Projeto = result.scalars().unique().one_or_none()
+    if usuario_logado:
+        query = select(Projeto).filter(Projeto.id == projeto_id)
+        result = db.execute(query)
+        projeto: Projeto = result.scalars().unique().one_or_none()
 
-    for member in projeto.membros:
-        if member.id == member_id:
-            db.delete(member)
-            db.commit()
-            return Response(detail="Membro removido com sucesso!", status_code=status.HTTP_204_NO_CONTENT)
-    
-    else:
-        raise HTTPException(detail="Membro não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
+        for member in projeto.membros:
+            if member.id == member_id:
+                db.delete(member)
+                db.commit()
+                return Response(detail="Membro removido com sucesso!", status_code=status.HTTP_204_NO_CONTENT)
+        
+        else:
+            raise HTTPException(detail="Membro não encontrado!", status_code=status.HTTP_404_NOT_FOUND)
         
 #DELETE laboratorio
 @router.delete('/{projeto_id}', status_code=status.HTTP_204_NO_CONTENT)
